+++ packages/pancake-uikit/src/components/BaseMenu/InlineMenu.tsx
@@ -1,15 +1,15 @@
 import React from "react";
 import { BoxProps } from "../Box";
 import BaseMenu from "./BaseMenu";
-import { MenuContainer } from "./styles";
+import { InlineMenuContainer } from "./styles";
 import { BaseMenuProps } from "./types";
 
-const Menu: React.FC<BaseMenuProps & BoxProps> = ({ children, component, isOpen = false, ...props }) => {
+const InlineMenu: React.FC<BaseMenuProps & BoxProps> = ({ children, component, isOpen = false, ...props }) => {
   return (
     <BaseMenu options={{ placement: "bottom" }} component={component} isOpen={isOpen}>
-      <MenuContainer {...props}>{children}</MenuContainer>
+      <InlineMenuContainer {...props}>{children}</InlineMenuContainer>
     </BaseMenu>
   );
 };
 
-export default Menu;
+export default InlineMenu;
