+++ packages/pancake-uikit/src/components/BaseMenu/index.tsx
@@ -1,5 +1,5 @@
 export { default as BaseMenu } from "./BaseMenu";
-export { default as Menu } from "./Menu";
+export { default as InlineMenu } from "./InlineMenu";
 export { default as SubMenu } from "./SubMenu";
 export * from "./styles";
 export type { MenuOptions, BaseMenuProps } from "./types";
