+++ packages/pancake-uikit/src/components/BaseMenu/styles.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import { Box, Flex } from "../Box";
 
-export const MenuContainer = styled(Box)`
+export const InlineMenuContainer = styled(Box)`
   background-color: ${({ theme }) => theme.colors.backgroundAlt};
   border-radius: ${({ theme }) => theme.radii.card};
   box-shadow: ${({ theme }) => theme.shadows.tooltip};
